# src/main/java/com/vendaprodutos/model/Produto.java
package com.vendaprodutos.model;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;

import com.vendaprodutos.model.dto.ProdutoDto;

@Entity
public class Produto {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long id;

	private String descricao;
	private Double precoUnitario;
	private String status;

	public static Produto from(ProdutoDto produtoDto) {
		Produto produto = new Produto();
		produto.setDescricao(produtoDto.getDescricao());
		produto.setPrecoUnitario(produtoDto.getPrecoUnitario());
		return produto;
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getDescricao() {
		return descricao;
	}

	public void setDescricao(String descricao) {
		this.descricao = descricao;
	}

	public Double getPrecoUnitario() {
		return precoUnitario;
	}

	public void setPrecoUnitario(Double precoUnitario) {
		this.precoUnitario = precoUnitario;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}

}